connection 结构调整
