+++ src/main/java/com/feeyo/kafka/net/backend/KafkaBackendConnectionFactory.java
@@ -18,11 +18,12 @@ public class KafkaBackendConnectionFactory implements BackendConnectionFactory {
 
 		String host = physicalNode.getHost();
 		int port = physicalNode.getPort();
+		boolean isZeroCopy = physicalNode.isZeroCopy();
 		
 		SocketChannel channel = SocketChannel.open();
 		channel.configureBlocking(false);
 
-		KafkaBackendConnection c = new KafkaBackendConnection( channel );
+		KafkaBackendConnection c = new KafkaBackendConnection(isZeroCopy, channel );
 		NetSystem.getInstance().setSocketParams(c, false);
 
 		// 设置NIOHandlers
