+++ src/main/java/com/feeyo/redis/net/backend/RedisBackendConnection.java
@@ -16,8 +16,8 @@ public class RedisBackendConnection extends BackendConnection {
 
 	private volatile int db = 0;				//REDIS select database, default 0
     
-    public RedisBackendConnection(SocketChannel channel) {
-		super(channel);
+    public RedisBackendConnection(boolean isZeroCopy, SocketChannel channel) {
+		super(isZeroCopy, channel);
 	}
 
 	public void setDb(int db) {
