+++ src/main/java/com/feeyo/redis/net/backend/RedisBackendConnectionFactory.java
@@ -11,16 +11,17 @@ import com.feeyo.redis.nio.NetSystem;
 public class RedisBackendConnectionFactory implements BackendConnectionFactory {
 
 	@Override
-	public BackendConnection make(PhysicalNode physicalNode, 
+	public BackendConnection make(PhysicalNode physicalNode,
 			BackendCallback callback, Object attachement) throws IOException {
 		
 		String host = physicalNode.getHost();
 		int port = physicalNode.getPort();
+		boolean isZeroCopy = physicalNode.isZeroCopy();
 		
 		SocketChannel channel = SocketChannel.open();
 		channel.configureBlocking(false);
 
-		RedisBackendConnection c = new RedisBackendConnection( channel );
+		RedisBackendConnection c = new RedisBackendConnection( isZeroCopy, channel );
 		NetSystem.getInstance().setSocketParams(c, false);
 
 		// 设置NIOHandlers
