+++ src/main/java/com/feeyo/redis/net/backend/pool/cluster/RedisClusterPool.java
@@ -236,10 +236,12 @@ public class RedisClusterPool extends AbstractPool {
 				String name = poolCfg.getName();
 				int minCon = poolCfg.getMinCon();
 				int maxCon = poolCfg.getMaxCon();
+				boolean isZeroCopy = poolCfg.isZeroCopy();
+				
 				String host = clusterNode.getHost();
 				int port = clusterNode.getPort();
 				
-				PhysicalNode physicalNode = new PhysicalNode(backendConFactory, type, name, minCon, maxCon, host, port );
+				PhysicalNode physicalNode = new PhysicalNode(backendConFactory, type, name, minCon, maxCon, host, port, isZeroCopy );
 				physicalNode.initConnections();
 				clusterNode.setPhysicalNode(physicalNode);				
 			}			
@@ -454,10 +456,12 @@ public class RedisClusterPool extends AbstractPool {
 						String name = poolCfg.getName();
 						int minCon = poolCfg.getMinCon();
 						int maxCon = poolCfg.getMaxCon();
+						boolean isZeroCopy = poolCfg.isZeroCopy();
+						
 						String host = clusterNode.getHost();
 						int port = clusterNode.getPort();
 							
-						PhysicalNode physicalNode = new PhysicalNode(backendConFactory, type, name, minCon, maxCon, host, port );
+						PhysicalNode physicalNode = new PhysicalNode(backendConFactory, type, name, minCon, maxCon, host, port, isZeroCopy );
 						physicalNode.initConnections();
 						clusterNode.setPhysicalNode( physicalNode );
 					}
