+++ src/main/java/com/feeyo/redis/net/backend/pool/xcluster/XClusterPool.java
@@ -37,7 +37,8 @@ public class XClusterPool extends AbstractPool{
         String name = poolCfg.getName();
         int minCon = poolCfg.getMinCon();
         int maxCon = poolCfg.getMaxCon();
-
+        boolean isZeroCopy = poolCfg.isZeroCopy();
+        
         for (String nodeStr : poolCfg.getNodes()) {
             
             String[] attrs = nodeStr.split(":");
@@ -48,7 +49,7 @@ public class XClusterPool extends AbstractPool{
             xNode.setSuffix( attrs[2] );
 
             PhysicalNode physicalNode = new PhysicalNode(backendConFactory, 
-            		type, name, minCon, maxCon, xNode.getIp(), xNode.getPort());
+            		type, name, minCon, maxCon, xNode.getIp(), xNode.getPort(), isZeroCopy);
             physicalNode.initConnections();
             xNode.setPhysicalNode(physicalNode);
