--- src/main/java/com/feeyo/redis/nio/test/ZeroCopyConnection.java
@@ -1,13 +1,0 @@
-package com.feeyo.redis.nio.test;
-
-import java.nio.channels.SocketChannel;
-
-import com.feeyo.redis.nio.AbstractZeroCopyConnection;
-
-public class ZeroCopyConnection extends AbstractZeroCopyConnection {
-
-	public ZeroCopyConnection(SocketChannel channel) {
-		super(channel);
-	}
-
-}
